Refactor: update error handling for command handlers and session (#1)
session.py
- add centralized error handling for command handlers via try/catch blocks with appropriate messages
- add try/catch to class variable initialization to handle errors doing auth, raise exception to main
- move except KeyboardInterrupt to top level (main.py) for clarity

main.py
- add try/catch to help program exit gracefully on exceptions

commands/*
- remove print statement from command handlers for failed commands, session.py handles with except blocks for better consistency
- raise FileNotFoundError instead of printing "No such item: ..."

+++ src/commands/upload.py
@@ -1,10 +1,19 @@
+import os
+from drive_tree.drive_node import DriveNode
+
+
 def handle_upload(session, file_path):
     client = session.client
     curr_dir = session.cwd
 
+    if not os.path.exists(file_path):
+        raise FileNotFoundError(file_path)
+
     uploaded_item = client.upload_file(file_path, curr_dir.id)
-    if not uploaded_item:
-        print(f"Failed to upload: {uploaded_item.get('name')}")
-        return
-    else:
-        print(f"Successfully uploaded: {uploaded_item.get('name')}")
+    uploaded_item_node = DriveNode(
+        uploaded_item.get("id"),
+        uploaded_item.get("name"),
+        uploaded_item.get("mimeType"),
+    )
+    curr_dir.add_child(uploaded_item_node)
+    print(f"Successfully uploaded: {uploaded_item.get('name')}")
